refactor: 修改参数类型
