+++ packages/layers/src/imagetile/index.ts
@@ -1,20 +1,19 @@
 import LayerGroup from '../core/LayerGroup';
 import ImageTileModels, { ImageTileModelType } from './models/index';
-interface IImageLayerStyleOptions {
-  opacity: number;
-}
-export default class ImageTileLayer extends LayerGroup<
-  IImageLayerStyleOptions
-> {
+
+export default class ImageTileLayer extends LayerGroup<ITileLayerStyleOptions> {
   public type: string = 'ImageTileLayer';
+
   public buildModels() {
     const modelType = this.getModelType();
     this.layerModel = new ImageTileModels[modelType](this);
     this.models = this.layerModel.initModels();
   }
+
   public rebuildModels() {
     this.models = this.layerModel.buildModels();
   }
+
   protected getConfigSchema() {
     return {
       properties: {
@@ -26,6 +25,7 @@ export default class ImageTileLayer extends LayerGroup<
       },
     };
   }
+
   protected getDefaultConfig() {
     const type = this.getModelType();
     const defaultConfig = {
