+++ packages/source/src/index.ts
@@ -4,7 +4,7 @@ import geojson from './parser/geojson';
 import image from './parser/image';
 import json, { defaultData, defaultParser } from './parser/json';
 import raster from './parser/raster';
-import tile from './parser/rasterTile';
+import rasterTile from './parser/raster-tile';
 import Source from './source';
 import { cluster } from './transform/cluster';
 import { filter } from './transform/filter';
@@ -12,7 +12,8 @@ import { aggregatorToGrid } from './transform/grid';
 import { pointToHexbin } from './transform/hexagon';
 import { join } from './transform/join';
 import { map } from './transform/map';
-registerParser('rasterTile', tile);
+
+registerParser('rasterTile', rasterTile);
 registerParser('geojson', geojson);
 registerParser('image', image);
 registerParser('csv', csv);
@@ -24,6 +25,7 @@ registerTransform('join', join);
 registerTransform('map', map);
 registerTransform('grid', aggregatorToGrid);
 registerTransform('hexagon', pointToHexbin);
+
 export {
   getTransform,
   registerTransform,
@@ -35,4 +37,5 @@ export * from './interface';
 
 export const DEFAULT_DATA = defaultData;
 export const DEFAULT_PARSER = defaultParser;
+
 export default Source;
